Converted pounds from int to decimal, need to do it everywhere

+++ CCIA/Areas/Admin/Models/Reports/OECDReport.cs
@@ -13,7 +13,7 @@ namespace CCIA.Models
        public string OECDClass { get; set; }
 
        public string report_group { get; set; }
-       public int pounds_oecd { get; set; }
+       public decimal pounds_oecd { get; set; }
        
        public int number_certificates{ get; set; }
